requirements.txt
asgiref==3.2.3
asn1crypto==1.4.0
cffi==1.14.5
cryptography==2.1.4
Django==3.0.1
django-debug-toolbar==2.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework==3.12.4
idna==3.1
pycparser==2.20
python-decouple==3.3
pytz==2019.3
six==1.15.0
sqlparse==0.3.0
